requirements.txt
absl-py==2.1.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flatbuffers==25.2.10
fonttools==4.56.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jax==0.5.1
jaxlib==0.5.1
kiwisolver==1.4.8
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scipy==1.15.2
sentencepiece==0.2.0
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sounddevice==0.5.1
sqlparse==0.5.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
zope.interface==7.2
